+++ bin/cli.js
@@ -1,22 +1,23 @@
 #!/usr/bin/env node
 
-import fs from "fs"
-import path from "path"
+import { existsSync, mkdirSync, readdirSync } from "fs"
+import { join } from "path"
 import { error } from "../src/functions.js"
 
-import Command from "../src/command.js"
-import index from "../src/index.js"
+import { default as Command } from "../src/command.js"
+import { default as index } from "../src/index.js"
 
 let _argv = process.argv.slice(2)
 let rootPath = process.cwd()
 
-if (_argv.length) rootPath = path.join(rootPath, _argv[0])
-if (!fs.existsSync(rootPath) || !fs.readdirSync(rootPath).length) {
+if (_argv.length) rootPath = join(rootPath, _argv[0])
+if (!existsSync(rootPath) || !readdirSync(rootPath).length) {
 	let command = new Command
 
-	command.get()
+	command
+		.get()
 		.then(( answers ) => {
-			if (!fs.existsSync(rootPath)) fs.mkdirSync(rootPath, { recursive: true })
+			if (!existsSync(rootPath)) mkdirSync(rootPath, { recursive: true })
 			index(answers, rootPath)
 		})
 } else {