+++ src/create/svelte.js
@@ -1,5 +1,5 @@
-import * as importSvelte from "./svelte/svelte.js"
+import * as impSvelte from "./svelte/svelte.js"
 
-export const svelte = ( files, options ) => importSvelte[options.transpiler](files, options)
+export const svelte = ( files, options ) => impSvelte[options.transpiler](files, options)
 
 export { default as sveltekit } from "./svelte/sveltekit.js"